# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/repository/": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Retrieve all repositories from the database",
        "description": "This endpoint fetches a list of all repositories in the database.",
        "responses": {
          "200": {
            "description": "List of all repositories.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 2
                  },
                  "name": {
                    "type": "string",
                    "example": "Repo 2"
                  },
                  "description": {
                    "type": "string",
                    "example": "Repository 2 description"
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 3
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/repository/name/{name}": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Retrieve repository by name",
        "description": "This endpoint fetches repositories matching the given name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Name of the repository to search for"
          }
        ],
        "responses": {
          "200": {
            "description": "List of repositories matching the name.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Repo 1"
                  },
                  "description": {
                    "type": "string",
                    "example": "Repository 1 description"
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No repository found with the specified name.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such repository with name: {name}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/repository/cp-id/{contactPersonId}": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Retrieve repositories by contact person ID",
        "description": "This endpoint fetches repositories associated with the given contact person ID.",
        "parameters": [
          {
            "name": "contactPersonId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Contact person ID to filter repositories by."
          }
        ],
        "responses": {
          "200": {
            "description": "List of repositories associated with the contact person ID.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Repo 1"
                  },
                  "description": {
                    "type": "string",
                    "example": "Repository 1 description"
                  },
                  "contactPersonId": {
                    "type": "number",
                    "example": 2
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No repositories found with the specified contact person ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such repository with contact person ID: {contactPersonId}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/repository/{id}": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Retrieve repository by ID",
        "description": "This endpoint fetches a repository based on its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the repository to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Repository found by ID.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "Repo 1"
                },
                "description": {
                  "type": "string",
                  "example": "Repository 1 description"
                },
                "contactPersonId": {
                  "type": "number",
                  "example": 2
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No repository found with the specified ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such repository with id: {id}"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}